+++ CHANGELOG.md
@@ -182,6 +182,11 @@ This version of std++ newly supports Rocq 9.1.
   for `gmultiset` operations, `elem_of`, `subseteq` and `multiplicity`.
   (by Egor Namakonov)
 - Extend the `multiset_solver` tactic to support the `filter` operation.
+- Remove the dependency of the class `TopSet` on `Set_` (and hence the
+  dependency on `Intersection` and `Difference`). (by Simcha van Collem)
+- Add the type `topGset` of finite sets with a top element. (by Simcha van
+  Collem)
+- Remove superfluous `Infinite` premise from `gset_to_coGset_finite`.
 
 **Changes in `stdpp_bitvector`:**